Rename lpips.py to perceptualsimilarity.py
